+++ ScoreBoard/Score.html
@@ -62,6 +62,7 @@
             &#169 2019 Emil,Lydia,Julia,Peter
         </footer>
     </div>
+    <script src="/ScoreBoard/highscoreScript.js" type="module"></script>
 </body>
 
 </html>
