+++ SnakeGame/game.html
@@ -3,7 +3,7 @@
 <html lang='en' xmlns='http://www.w3.org/1999/xhtml'>
 <head>
 <meta charset='utf-8' />
-<link rel='stylesheet' type='text/css' href='style.css' />
+<link rel='stylesheet' type='text/css' href='/SnakeGame/style.css' />
 <title></title>
 </head>
 <body>
@@ -40,6 +40,6 @@
         </div>
         
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-<script type='module' src='script.js' charset='utf-8'></script>
+<script type='module' src='/SnakeGame/script.js' charset='utf-8'></script>
 </body>
 </html>
