+++ mainpage/main.html
@@ -38,7 +38,7 @@
     <div class="leftmidcircle"></div>
     </div>
     <ul>
-        <li><a href="/game.html">Start</a></li>
+        <li><a href="/SnakeGame/game.html">Start</a></li>
         <li><a href="/Instruction/Instruction.html">How to play</a></li>
         <li><a href="/ScoreBoard/Score.html">Score board</a></li>
     </ul>
